+++ src/jk_exceptionhelper/StackTraceItem.py
@@ -1,14 +1,19 @@
 
 
-# import collections
+__all__ = (
+	"ExceptionObject",
+)
 
-# StackTraceItem = collections.namedtuple("StackTraceItem", [
-#	"filePath", "lineNo", "callingScope", "sourceCodeLine"
-# ])
+
+
 
 
 class StackTraceItem(object):
-	
+
+	################################################################################################################################
+	## Constructor
+	################################################################################################################################
+
 	def __init__(self, filePath:str, lineNo:int, callingScope:str, sourceCodeLine:str):
 		assert isinstance(filePath, str)
 		assert isinstance(lineNo, int)
@@ -20,6 +25,18 @@ class StackTraceItem(object):
 		self.callingScope = callingScope
 		self.sourceCodeLine = sourceCodeLine
 	#
+
+	################################################################################################################################
+	## Public Properties
+	################################################################################################################################
+
+	################################################################################################################################
+	## Helper Methods
+	################################################################################################################################
+
+	################################################################################################################################
+	## Public Methods
+	################################################################################################################################
 
 	def __str__(self):
 		return self.filePath + ":" + str(self.lineNo) + " :: " + repr(self.sourceCodeLine)
@@ -38,8 +55,27 @@ class StackTraceItem(object):
 		}
 	#
 
+	def __eq__(self, other):
+		if isinstance(other, StackTraceItem):
+			return (other.filePath == self.filePath) and (other.lineNo == self.lineNo)
+		else:
+			return False
+	#
+
+	def __ne__(self, other):
+		if isinstance(other, StackTraceItem):
+			return (other.filePath != self.filePath) or (other.lineNo != self.lineNo)
+		else:
+			return True
+	#
+
+	################################################################################################################################
+	## Public Static Methods
+	################################################################################################################################
+
 	@staticmethod
 	def fromJSON(data:dict):
+		assert isinstance(data, dict)
 		return StackTraceItem(
 			data["file"],
 			data["line"],