+++ src/jk_exceptionhelper/__init__.py
@@ -2,11 +2,12 @@
 
 
 __author__ = "Jürgen Knauth"
-__version__ = "0.2023.7.1"
+__version__ = "0.2025.5.17"
 
 
 
 from .StackTraceItem import StackTraceItem
+from .StackTraceProcessor import StackTraceProcessor
 from .ExceptionObject import ExceptionObject
 
 from .exception_helper import analyseException