+++ HISTORY.md
@@ -1,8 +1,15 @@
-* 2021-12-14
+* 2021-12-14:
 	* Refactoring
 	* Extended string matching to non-POSIX paths
 
-* 2923-07-01
+* 2023-07-01:
 	* Refactoring
 	* Improved: Processing nested exceptions
+
+* 2024-01-03:
+	* Migrated to TOML file
+	* Improved: Processing stack traces and nested exceptions
+
+* 2025-05-17:
+	* Improved: Support for extra key value pairs via 'extraValues'